+++ Reading List App/Reading List App/AppDelegate.swift
@@ -23,7 +23,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         // Initialize Parse.
         Parse.setApplicationId("FyE1vRjeMYXISsCcrmmcjN64Dkw0ozLpru6siAWH",
             clientKey: "hpD0Q322NhuppTDK2bbMoJtKJ9L0BULXOpW7zWBm")
-        PFTwitterUtils.initializeWithConsumerKey("YOUR CONSUMER KEY",  consumerSecret:"YOUR CONSUMER SECRET")
+        PFTwitterUtils.initializeWithConsumerKey("ASK_@ALEXBTLV_FOR_KEYS",  consumerSecret:"ASK_@ALEXBTLV_FOR_KEYS")
         
         // [Optional] Track statistics around application opens.
         PFAnalytics.trackAppOpenedWithLaunchOptions(launchOptions)
